Work so far, showing basis of the image gallery, no unit tests as yet.
Shows master/detail design (detail not yet implemented), with the
master connecting to flickr via a data controller and parsing the JSON
result to a table. The meta data are shown in the table. The intention
is to implement an asynchronous fetch of the images for display in the
table.